third_party/WebKit/LayoutTests/inspector/initial-modules-load-expected.txt

This test validates initial set of loaded modules.


Running: testInitialLoad
Loaded modules:
    bindings
    common
    components
    console_model
    dom_extension
    emulation
    extensions
    help
    host
    main
    mobile_throttling
    network_log
    persistence
    platform
    product_registry
    protocol
    sdk
    services
    text_utils
    ui
    workspace

Running: testCreateElementsPanel
Loaded modules:
    bindings
    color_picker
    common
    components
    console_model
    dom_extension
    elements
    emulation
    event_listeners
    extensions
    help
    host
    inline_editor
    main
    mobile_throttling
    network_log
    object_ui
    persistence
    platform
    product_registry
    protocol
    sdk
    services
    text_utils
    ui
    workspace

Running: testCreateNetworkPanel
Loaded modules:
    animation
    bindings
    cm
    color_picker
    common
    components
    console_model
    cookie_table
    data_grid
    diff
    dom_extension
    elements
    emulation
    event_listeners
    extensions
    formatter
    help
    host
    inline_editor
    main
    mobile_throttling
    network
    network_log
    network_priorities
    object_ui
    perf_ui
    persistence
    platform
    product_registry
    protocol
    sdk
    services
    source_frame
    text_editor
    text_utils
    ui
    workspace
    workspace_diff

Running: testShowSourcesPanel
Loaded modules:
    animation
    bindings
    cm
    color_picker
    common
    components
    console_model
    cookie_table
    data_grid
    diff
    dom_extension
    elements
    emulation
    event_listeners
    extensions
    formatter
    help
    host
    inline_editor
    main
    mobile_throttling
    network
    network_log
    network_priorities
    object_ui
    perf_ui
    persistence
    platform
    product_registry
    protocol
    quick_open
    sdk
    services
    snippets
    source_frame
    sources
    text_editor
    text_utils
    ui
    workspace
    workspace_diff

Running: testOpenUISourceCode
Loaded modules:
    animation
    bindings
    cm
    color_picker
    common
    components
    console_model
    cookie_table
    data_grid
    diff
    dom_extension
    elements
    emulation
    event_listeners
    extensions
    formatter
    help
    host
    inline_editor
    main
    mobile_throttling
    network
    network_log
    network_priorities
    object_ui
    perf_ui
    persistence
    platform
    product_registry
    protocol
    quick_open
    sdk
    services
    snippets
    source_frame
    sources
    text_editor
    text_utils
    ui
    workspace
    workspace_diff
